+++ src/services/data/documents.ts
@@ -150,17 +150,21 @@ export default (
       ).data;
     },
 
-    async unlinkGroups(schemaIdOrName, documentId, requestBody, options) {
-      const data = requestBody || {};
+    async unlinkGroups(schemaIdOrName, documentId, data: Array<string> | { groupIds?: Array<string>; }, options) {
+      const requestBody = Array.isArray(data) ? { groupIds: data } : data;
 
       return (
         await client.post(
           httpAuth,
           `/${schemaIdOrName}/documents/${documentId}/unlinkGroups`,
-          data,
+          requestBody,
           options
         )
       ).data;
+    },
+
+    async unlinkAllGroups(schemaIdOrName, documentId, options) {
+      return await this.unlinkGroups(schemaIdOrName, documentId, {}, options); // Empty object to remove all groups
     },
 
     async linkUsers(schemaIdOrName, documentId, requestBody, options) {
@@ -174,17 +178,21 @@ export default (
       ).data;
     },
 
-    async unlinkUsers(schemaIdOrName, documentId, requestBody, options) {
-      const data = requestBody || {};
+    async unlinkUsers(schemaIdOrName, documentId, data: Array<string> | { userIds?: Array<string>; }, options) {
+      const requestBody = Array.isArray(data) ? { userIds: data } : data;
 
       return (
         await client.post(
           httpAuth,
           `/${schemaIdOrName}/documents/${documentId}/unlinkUsers`,
-          data,
+          requestBody,
           options
         )
       ).data;
+    },
+
+    async unlinkAllUsers(schemaIdOrName, documentId, options) {
+      return await this.unlinkUsers(schemaIdOrName, documentId, {}, options); // Empty object to remove all users
     },
   };
 };